Move SaveSyncSharedPreferencesShouldVerifyDataSave to ConfigurationRegistry tests
