+++ android/engine/src/test/java/org/smartregister/fhircore/engine/configuration/ConfigurationRegistryTest.kt
@@ -25,6 +25,7 @@ import com.google.android.fhir.SearchResult
 import com.google.android.fhir.db.ResourceNotFoundException
 import com.google.android.fhir.logicalId
 import com.google.android.fhir.search.Search
+import com.google.common.reflect.TypeToken
 import com.google.gson.GsonBuilder
 import dagger.hilt.android.testing.HiltAndroidRule
 import dagger.hilt.android.testing.HiltAndroidTest
@@ -792,4 +793,28 @@ class ConfigurationRegistryTest : RobolectricTest() {
       Assert.assertEquals("composition-id-1", requestPathArgumentSlot.last().id)
       Assert.assertEquals(ResourceType.Composition, requestPathArgumentSlot.last().resourceType)
     }
+
+  @Test
+  fun testSaveSyncSharedPreferencesShouldVerifyDataSave() {
+    val resourceType =
+      listOf(ResourceType.Task, ResourceType.Patient, ResourceType.Task, ResourceType.Patient)
+
+    configRegistry.saveSyncSharedPreferences(resourceType)
+
+    val savedSyncResourcesResult =
+      configRegistry.sharedPreferencesHelper.read(
+        SharedPreferenceKey.REMOTE_SYNC_RESOURCES.name,
+        null,
+      )!!
+    val listResourceTypeToken = object : TypeToken<List<ResourceType>>() {}.type
+    val savedSyncResourceTypes: List<ResourceType> =
+      configRegistry.sharedPreferencesHelper.gson.fromJson(
+        savedSyncResourcesResult,
+        listResourceTypeToken,
+      )
+
+    Assert.assertEquals(2, savedSyncResourceTypes.size)
+    Assert.assertEquals(ResourceType.Task, savedSyncResourceTypes.first())
+    Assert.assertEquals(ResourceType.Patient, savedSyncResourceTypes.last())
+  }
 }
